+++ package.json
@@ -13,6 +13,7 @@
     "react": "^17.0.1",
     "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.0",
     "recharts": "^1.8.5",
     "web-vitals": "^0.2.4"
