+++ src/components/Nav.js
@@ -1,1 +1,11 @@
-//proba
+import React from 'react';
+
+const Nav = () => {
+   
+  
+  return (
+     <div>Hello World</div> 
+  );
+};
+
+export default Nav;
